requirements.txt
absl-py==1.3.0
accelerate==0.17.1
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bitsandbytes==0.37.2
bleach==4.1.0
blessings==1.7
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
clip==1.0
comet-ml==3.32.2
comm==0.1.2
configobj==5.0.8
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
dulwich==0.21.3
einops==0.6.0
entrypoints==0.4
everett==3.1.0
executing==0.8.3
fairscale==0.4.13
faiss==1.7.3
fastapi==0.95.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
flit-core==3.6.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==5.8
fvcore==0.1.5.post20221221
gitdb==4.0.10
gitpython==3.1.31
google-auth-oauthlib==0.4.4
google-auth==2.6.0
gpustat==0.6.0
gputil==1.4.0
gradio==3.23.0
grpcio==1.42.0
h11==0.14.0
hiq-python==1.1.9
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
hydra-core==1.3.2
idna==3.4
importlib-metadata==4.11.3
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.19.2
ipython-genutils==0.2.0
ipython==8.10.0
ipywidgets==7.6.5
jedi==0.18.1
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==5.2.0
jupyter-server==1.23.4
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.4
linkify-it-py==2.0.0
lxml==4.9.1
markdown-it-py==2.2.0
markdown==3.4.1
markupsafe==2.1.1
matplotlib-inline==0.1.6
matplotlib==3.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook-shim==0.2.2
notebook==6.5.2
numpy==1.24.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.1
omegaconf==2.3.0
openai==0.27.4
opencv-python==4.7.0.72
orjson==3.8.8
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pip==22.3.1
platformdirs==2.5.2
portalocker==2.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-itree==0.0.18
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
pygments==2.11.2
pyjwt==2.4.0
pyllama==0.0.8
pyopenssl==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
pysocks==1.7.1
python-box==6.1.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2022.7.1
pyyaml==6.0
pyzmq==23.2.0
regex==2022.10.31
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
send2trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==65.6.3
simplejson==3.18.3
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
starlette==0.26.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.13.2
toolz==0.12.0
torch==1.12.1+cu113
torchmetrics==0.11.4
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
transformers==4.27.3
typing-extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
wandb==0.13.10
warmup-scheduler==0.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
websockets==10.4
werkzeug==2.2.2
wheel==0.38.4
widgetsnbextension==3.5.2
wrapt==1.14.1
wurlitzer==3.0.3
yacs==0.1.8
yarl==1.8.1
zipp==3.11.0